refactor: eliminate the need for client code to use DisplayHelp
Since we already know what gets passed in at parsing time, we may as
well print the help message right away, so that we can exit the
program for the user.

The point is precisely that: when -h/--help is requested, the program
should print the help message, then exit.